refactoring, removal of unused functions and pylint
